requirements.txt
requests
python-dotenv
typeguard
pyyaml
llama-cpp-python
huggingface-hub
groq
instructor
pydantic